pyproject.toml
[project]
name = "db-mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "mcp[cli]>=1.26.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
]
[project.scripts]
pg-mcp = "src.pg_mcp.server:mcp"

[tool.uv]
package = true